MAPI-9 Fixed the page logic issues
As Part of Automation Maintenance found the below issues due to platform upgradation and fixed the same.
My Calendar: While editing the meeting form, Daily to One time meeting some of the parameters like Timestamp, Start date, End date, Recurrence value in the response is as "No text to display" so there is service action failure.
My Calendar: While editing the attendee, Once the user click on the "can't find user ?", Attendee 1 [Enter Email address] (optional) text box poops up, with text -"No text to display" and it not editable. (IF-174 has been raised for this)
My office hours : The logged in user name is not dynamically visible in event title while creating/editing/viewing the office hours.
 

Fix for the issue:

My office hours: After analyzing able to understand that the existing setup is correct, tried deleting and adding again the filter condition is fixed the issue.
My Calendar: Updated the default value of the Text-Box from Blank-space to empty. Also fine tuned couple of page level logics to fix the issue.